refactor: unify icon mode detection and migrate icon registry to shared modules
Extract icon mode detection (AtomicBool, init_from_config, use_nerd_font)
into src/icon_mode.rs and icon registry (IconId, IconPair, get, get_for_mode)
into src/icons.rs — both accessible from the lib crate. SessionStatus now
delegates nerd_symbol() to the centralized registry via icon_id() mapping.
Removes duplicated AtomicBool in session/types.rs, eliminates set_ascii_icons(),
and consolidates startup init to a single call site.

Closes #307
Closes #308
