+++ src/lib.rs
@@ -1,4 +1,7 @@
 /// Library facade — exposes only self-contained modules for benchmarks.
+pub mod icon_mode;
+pub mod icons;
+
 #[path = "session"]
 pub mod session {
     pub mod parser;
