+++ src/main.rs
@@ -9,6 +9,8 @@ mod flags;
 mod gates;
 mod git;
 mod github;
+mod icon_mode;
+mod icons;
 mod models;
 mod modes;
 mod notifications;
